reorg internal types, remove extraneous imports
