+++ source/internal/index.d.ts
@@ -1,0 +1,8 @@
+export type * from './array';
+export type * from './characters';
+export type * from './keys';
+export type * from './numeric';
+export type * from './object';
+export type * from './string';
+export type * from './tuple';
+export type * from './type';
