+++ source/undefined-on-partial-deep.d.ts
@@ -1,5 +1,4 @@
 import type {BuiltIns} from './internal';
-import type {Merge} from './merge';
 
 /**
 Create a deep version of another type where all optional keys are set to also accept `undefined`.
